fix appearance caching lua errors
with patch 9.2.5 the APIs for `C_TransmogCollection.GetCategoryAppearances`, and `C_TransmogCollection.GetAppearanceSources` changed. The former no longer uses the exclusion parameter but rather a transmogLocation object (same for GetAppearanceSources). Said transmogLocation can be retrieved from `TransmogUtil.GetTransmogLocation(slotId, type, modfication)`. The fix currently ignores additional slots like the secondary shoulder slot, which shouldn't be an issue in this case. Unfortunately, there doesn't seem to be a way to map a transmog category to a inventory slot (atleast I didn't find one), so I had to add a table to map the categories to an inventory slot.